refactor: Изменить обработку ошибок
