+++ src/main/java/ru/yandex/practicum/filmorate/error/ErrorResponse.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.model;
+package ru.yandex.practicum.filmorate.error;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
