+++ src/main/java/ru/yandex/practicum/filmorate/error/Violation.java
@@ -1,0 +1,11 @@
+package ru.yandex.practicum.filmorate.error;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class Violation {
+    private final String fieldName;
+    private final String message;
+}
